pyproject.toml
[project]
name = "voiceforge"
version = "0.1.0a1"
description = "Local-first AI assistant for audio meetings: STT, diarization, RAG, LLM analysis"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    {name = "VoiceForge Contributors"},
]
keywords = ["ai", "meeting", "transcription", "diarization", "rag", "llm", "local-first"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "typer>=0.12",
    "structlog>=24.0",
    "psutil>=5.9",
    "numpy>=1.26,<3",
    "faster-whisper>=1.1.0,<2.0",
    "dbus-fast>=2.0",
    "pyannote.audio==3.3.2",
    "matplotlib>=3.8",
    "keyring>=25.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
rag = [
    "onnxruntime>=1.18,<2",
    "sqlite-vec>=0.1.0",
    "pymupdf>=1.24",
    "tokenizers>=0.19",
    "huggingface-hub>=0.20",
    "watchdog>=4.0",
    "python-docx>=1.0",
]
llm = [
    "litellm>=1.50,<2",
    "instructor>=1.0,<2",
    "pydantic>=2.0",
]
pii = ["gliner>=0.2.20"]
all = ["voiceforge[rag,llm,pii]"]

[project.scripts]
voiceforge = "voiceforge.main:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "bandit>=1.9.3",
    "pip-audit>=2.10.0",
    "mypy>=1.19.1",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.ruff]
line-length = 130
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = [
    "RUF001", "RUF002", "RUF003",
    "SIM102", "SIM108",
    "UP047",
    "E501",
]

[tool.ruff.lint.isort]
known-first-party = ["voiceforge"]

[tool.ruff.lint.per-file-ignores]
"src/voiceforge/stt/diarizer.py" = ["E402"]
"src/voiceforge/main.py" = ["B904", "SIM105"]
"src/voiceforge/i18n/__init__.py" = ["SIM105"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/voiceforge"]
branch = true
omit = [
    "*/tests/*",
    "src/voiceforge/__main__.py",
    "src/voiceforge/main.py",
    "src/voiceforge/core/daemon.py",
    "src/voiceforge/audio/smart_trigger.py",
    "src/voiceforge/core/model_manager.py",
    "src/voiceforge/stt/streaming.py",
]

[tool.coverage.report]
fail_under = 80
exclude_lines = ["pragma: no cover", "def __repr__", "raise NotImplementedError"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voiceforge"]
